chore: More icon cleanup
