+++ package.json
@@ -29,7 +29,6 @@
     "@reach/router": "^1.3.3",
     "@reach/skip-nav": "^0.9.0",
     "@reach/tooltip": "^0.10.5",
-    "@reach/visually-hidden": "^0.9.0",
     "algoliasearch": "^4.1.0",
     "classnames": "^2.2.6",
     "copy-to-clipboard": "^3.3.1",
