+++ src/components/layout/header/development-warning.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { StaticQuery, graphql } from 'gatsby'
-import Alert from '~components/utils/alert'
 import developmentWarningStyle from './development-warning.module.scss'
 
 export default () => (
@@ -20,11 +19,9 @@ export default () => (
           <div
             className={`development-warning ${developmentWarningStyle.developmentWarning}`}
           >
-            <Alert>
-              This is a test version of The COVID Tracking Project website. Data
-              may be inaccurate.{' '}
-              <a href="http://covidtracking.com/">Visit the live site</a>
-            </Alert>
+            This is a test version of The COVID Tracking Project website. Data
+            may be inaccurate.{' '}
+            <a href="http://covidtracking.com/">Visit the live site</a>
           </div>
         )}
       </>
