+++ src/components/pages/data/summary-charts.js
@@ -15,7 +15,7 @@ import { parseDate, formatDate } from '~utilities/visualization'
 import { Row, Col } from '~components/common/grid'
 import Toggle from '~components/common/toggle'
 import ContentfulContent from '~components/common/contentful-content'
-import { AlertInfobox } from '~components/common/infobox'
+import Alert from '~components/common/alert'
 import { FieldName } from '~components/utils/field-name'
 import { ReactComponent as CtpLogo } from '~images/project-logo.svg'
 import colors from '~scss/colors.module.scss'
@@ -82,8 +82,8 @@ const getDataForField = (data, field) => {
 }
 
 const ChartAlert = ({ message }) => (
-  <div className={styles.alertInfoboxContainer}>
-    <AlertInfobox header={message} fullSize />
+  <div className={styles.AlertContainer}>
+    <Alert header={message} fullSize />
   </div>
 )
