--- src/components/utils/alert.js
@@ -1,4 +1,0 @@
-import React from 'react'
-import Alert from '@reach/alert'
-
-export default ({ children }) => <Alert>{children}</Alert>
