+++ src/stories/2-components/10-Accessibility.stories.js
@@ -1,24 +1,9 @@
 import React from 'react'
-import VisuallyHidden from '@reach/visually-hidden'
-import Alert from '~components/utils/alert'
 import SkipNavigation from '~components/utils/skip-navigation'
 
 export default {
   title: 'Accessibility utilities',
 }
-
-export const visuallyHideComponent = () => (
-  <p>
-    This is a paragraph with some hidden text{' '}
-    <a href="/">
-      in this link <VisuallyHidden>ta da!</VisuallyHidden>
-    </a>
-  </p>
-)
-
-export const alert = () => (
-  <Alert>Woah, you should pay attention to this!</Alert>
-)
 
 export const skipNavigation = () => <SkipNavigation />
